Split document service factory
