+++ packages/tools/webpack-fluid-loader/src/getDocumentServiceFactory.ts
@@ -3,16 +3,18 @@
  * Licensed under the MIT License.
  */
 
-import { ILocalDeltaConnectionServer, LocalDeltaConnectionServer } from "@fluidframework/server-local-server";
-import { MultiDocumentServiceFactory } from "@fluidframework/driver-utils";
+import { assert } from "@fluidframework/common-utils";
+import { IDocumentServiceFactory } from "@fluidframework/driver-definitions";
 import { LocalDocumentServiceFactory, LocalSessionStorageDbFactory } from "@fluidframework/local-driver";
 import { OdspDocumentServiceFactory } from "@fluidframework/odsp-driver";
 import { HostStoragePolicy, IPersistedCache } from "@fluidframework/odsp-driver-definitions";
 import { RouterliciousDocumentServiceFactory } from "@fluidframework/routerlicious-driver";
+import { ILocalDeltaConnectionServer, LocalDeltaConnectionServer } from "@fluidframework/server-local-server";
 import { getRandomName } from "@fluidframework/server-services-client";
 import { InsecureTokenProvider } from "@fluidframework/test-runtime-utils";
-import { assert } from "@fluidframework/common-utils";
+
 import { v4 as uuid } from "uuid";
+
 import { IDevServerUser, IRouterliciousRouteOptions, RouteOptions } from "./loader";
 
 export const deltaConns = new Map<string, ILocalDeltaConnectionServer>();
@@ -22,7 +24,7 @@ export function getDocumentServiceFactory(
     options: RouteOptions,
     odspPersistantCache?: IPersistedCache,
     odspHostStoragePolicy?: HostStoragePolicy,
-) {
+): IDocumentServiceFactory {
     const deltaConn = deltaConns.get(documentId) ??
         LocalDeltaConnectionServer.create(new LocalSessionStorageDbFactory(documentId));
     deltaConns.set(documentId, deltaConn);
@@ -48,22 +50,31 @@ export function getDocumentServiceFactory(
             getUser());
     }
 
-    return MultiDocumentServiceFactory.create([
-        new LocalDocumentServiceFactory(deltaConn),
-        // TODO: web socket token
-        new OdspDocumentServiceFactory(
-            async () => options.mode === "spo" || options.mode === "spo-df" ? (options.odspAccessToken ?? null) : null,
-            async () => options.mode === "spo" || options.mode === "spo-df" ? (options.pushAccessToken ?? null) : null,
-            odspPersistantCache,
-            odspHostStoragePolicy,
-        ),
-        new RouterliciousDocumentServiceFactory(
-            routerliciousTokenProvider,
-            {
-                enableWholeSummaryUpload: options.mode === "r11s" || options.mode === "docker"
-                    ? options.enableWholeSummaryUpload
-                    : undefined,
-                enableDiscovery: options.mode === "r11s" && options.discoveryEndpoint !== undefined,
-            }),
-    ]);
+    switch (options.mode) {
+        case "docker":
+        case "r11s":
+        case "tinylicious":
+            return new RouterliciousDocumentServiceFactory(
+                routerliciousTokenProvider,
+                {
+                    enableWholeSummaryUpload: options.mode === "r11s" || options.mode === "docker"
+                        ? options.enableWholeSummaryUpload
+                        : undefined,
+                    enableDiscovery: options.mode === "r11s" && options.discoveryEndpoint !== undefined,
+                },
+            );
+
+        case "spo":
+        case "spo-df":
+            // TODO: web socket token
+            return new OdspDocumentServiceFactory(
+                async () => options.odspAccessToken ?? null,
+                async () => options.pushAccessToken ?? null,
+                odspPersistantCache,
+                odspHostStoragePolicy,
+            );
+
+        default: // Local
+            return new LocalDocumentServiceFactory(deltaConn);
+    }
 }
