+++ packages/tools/webpack-fluid-loader/src/routes.ts
@@ -20,7 +20,7 @@ import { IOdspTokens, getServer } from "@fluidframework/odsp-doclib-utils";
 import Axios from "axios";
 import { RouteOptions } from "./loader";
 import { createManifestResponse } from "./bohemiaIntercept";
-import { tinyliciousUrls } from "./multiResolver";
+import { tinyliciousUrls } from "./getUrlResolver";
 
 const tokenManager = new OdspTokenManager(odspTokensCache);
 let odspAuthStage = 0;
